Added date check to assingment upload

+++ class-management-api/routes/assignments.js
@@ -70,7 +70,10 @@ router.post('/:id/upload', auth, upload.single('file'), async (req, res) => {
         const userCourses = await req.user.courses();
         const assignmentCourse = await assignment.course();
         if (!userCourses.map(c => c._id.toString()).includes(assignmentCourse._id.toString())) {
-            return res.status(400).json({ message: 'Unable to upload item / not enrolled to course' });
+            return res.status(400).json({ message: 'Unable to upload item' });
+        }
+        if (assignment.dueDate.getTime() < new Date().getTime()) {
+            return res.status(400).json({ message: 'Time for upload expired' });
         }
         await Upload.deleteMany({ uploadedBy: req.user._id, owner: assignment._id.toString() });
         const upload = await Upload.createUpload(req.file, assignment._id.toString(), req.user._id);
